Moved all classes from util
