+++ modules/client/build.gradle
@@ -1,5 +1,6 @@
 plugins {
     id "org.jlleitschuh.gradle.ktlint"
+    id 'com.github.johnrengelman.shadow' version '6.1.0'
 }
 
 group 'jp.co.soramitsu'
