+++ modules/client/src/main/kotlin/jp/co/soramitsu/iroha2/Crypto.kt
@@ -1,9 +1,5 @@
-package jp.co.soramitsu.iroha2.utils
+package jp.co.soramitsu.iroha2
 
-import jp.co.soramitsu.iroha2.generated.datamodel.transaction.Payload
-import jp.co.soramitsu.iroha2.generated.datamodel.transaction.VersionedTransaction
-import jp.co.soramitsu.iroha2.utils.DigestFunction.Ed25519
-import net.i2p.crypto.eddsa.EdDSAEngine
 import net.i2p.crypto.eddsa.EdDSAPrivateKey
 import net.i2p.crypto.eddsa.EdDSAPublicKey
 import net.i2p.crypto.eddsa.spec.EdDSANamedCurveSpec
@@ -11,14 +7,9 @@ import net.i2p.crypto.eddsa.spec.EdDSANamedCurveTable
 import net.i2p.crypto.eddsa.spec.EdDSAParameterSpec
 import net.i2p.crypto.eddsa.spec.EdDSAPrivateKeySpec
 import net.i2p.crypto.eddsa.spec.EdDSAPublicKeySpec
-import org.bouncycastle.jcajce.provider.digest.Blake2b
 import org.bouncycastle.util.encoders.Hex
 import java.security.KeyPair
-import java.security.MessageDigest
-import java.security.PrivateKey
-import java.security.PublicKey
 import java.security.SecureRandom
-import jp.co.soramitsu.iroha2.generated.crypto.PublicKey as IrohaPublicKey
 
 val DEFAULT_SPEC: EdDSANamedCurveSpec = EdDSANamedCurveTable.getByName(EdDSANamedCurveTable.ED_25519)
 
@@ -38,36 +29,6 @@ fun generateKeyPair(spec: EdDSAParameterSpec = DEFAULT_SPEC): KeyPair {
         PrivateKeyWrapper(privKey)
     )
 }
-
-fun PublicKey.toIrohaPublicKey(): IrohaPublicKey {
-    return IrohaPublicKey(Ed25519.hashFunName, this.encoded)
-}
-
-/**
- * Sign the message by given private key
- *
- * Note: the message must not be prehashed
- */
-fun PrivateKey.sign(message: ByteArray): ByteArray {
-    val sgr = EdDSAEngine(MessageDigest.getInstance(DEFAULT_SPEC.hashAlgorithm))
-    sgr.initSign(this)
-    sgr.update(message.hash())
-    return sgr.sign()
-}
-
-/**
- * Verify the signature against the message and public key
- *
- * Note: the message must not be prehashed
- */
-fun PublicKey.verify(signature: ByteArray, message: ByteArray): Boolean {
-    val sgr = EdDSAEngine(MessageDigest.getInstance(DEFAULT_SPEC.hashAlgorithm))
-    sgr.initVerify(this)
-    sgr.update(message.hash())
-    return sgr.verify(signature)
-}
-
-fun ByteArray.hash(): ByteArray = Blake2b.Blake2b256().digest(this)
 
 fun keyPairFromHex(publicKeyHex: String, privateKeyHex: String, spec: EdDSAParameterSpec = DEFAULT_SPEC) =
     KeyPair(publicKeyFromHex(publicKeyHex, spec), privateKeyFromHex(privateKeyHex, spec))
@@ -89,14 +50,3 @@ class PrivateKeyWrapper(privKeySpec: EdDSAPrivateKeySpec) : EdDSAPrivateKey(priv
     override fun getFormat() = "RAW"
 }
 
-fun VersionedTransaction.V1.hash(): ByteArray {
-    return this._VersionedTransactionV1
-        .transaction
-        .payload
-        .encode(Payload)
-        .hash()
-}
-
-fun VersionedTransaction.hash() = when (this) {
-    is VersionedTransaction.V1 -> this.hash()
-}
