+++ modules/client/src/main/kotlin/jp/co/soramitsu/iroha2/TransactionBuilder.kt
@@ -8,9 +8,6 @@ import jp.co.soramitsu.iroha2.generated.datamodel.transaction.Payload
 import jp.co.soramitsu.iroha2.generated.datamodel.transaction.Transaction
 import jp.co.soramitsu.iroha2.generated.datamodel.transaction.VersionedTransaction
 import jp.co.soramitsu.iroha2.generated.datamodel.transaction._VersionedTransactionV1
-import jp.co.soramitsu.iroha2.utils.encode
-import jp.co.soramitsu.iroha2.utils.sign
-import jp.co.soramitsu.iroha2.utils.toIrohaPublicKey
 import java.security.KeyPair
 import java.time.Duration
 import java.time.Instant
